# feat: Claude Cookbook Patterns + AI Agent Backtesting Framework