--- a/.env.production
+++ b/.env.production
@@ -1,0 +1,17 @@
+# Phase 1 Feature Flags - Production Environment
+# ALL FLAGS DISABLED BY DEFAULT FOR SAFETY
+
+# Foundation Flags
+VITE_NEW_IA=false
+VITE_ENABLE_SELECTOR_KPIS=false
+
+# Build Flags
+VITE_ENABLE_MODELING_WIZARD=false
+VITE_ENABLE_OPERATIONS_HUB=false
+
+# Polish Flags
+VITE_ENABLE_LP_REPORTING=false
+
+# ⚠️ IMPORTANT: Only enable flags after thorough testing in preview deployment
+# To enable features, set flag to "true" in Vercel environment variables
+# DO NOT commit with flags=true to this file