--- a/Default Parameters/src/styles/brand-tokens.css
+++ b/Default Parameters/src/styles/brand-tokens.css
@@ -1,0 +1,12 @@
+
+@import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&family=Poppins:wght@400;500;600&display=swap');
+:root { --font-sans-1: 'Inter', system-ui; --font-sans-2: 'Poppins', system-ui; --color-ink:#292929; --color-paper:#fff; --color-mist:#F2F2F2; --color-sand:#E0D8D1; --bg-app:var(--color-mist); --bg-card:var(--color-paper); --border-muted:var(--color-sand); --text-primary:var(--color-ink); --text-muted:#5B5B5B; --radius-sm:6px; --radius-md:10px; --radius-lg:14px; --shadow-card:0 2px 12px rgba(0,0,0,.06); }
+html,body{font-family:var(--font-sans-1);color:var(--text-primary);background:var(--bg-app);}
+.card{background:var(--bg-card);border:1px solid var(--border-muted);border-radius:var(--radius-lg);box-shadow:var(--shadow-card);}
+.kpi-grid{display:grid;grid-template-columns:repeat(12,minmax(0,1fr));gap:12px;}
+.kpi{grid-column:span 3;padding:14px 16px;}
+.kpi .label{font:500 .8rem/1 var(--font-sans-2);color:var(--text-muted);}
+.kpi .value{font:600 1.3rem/1.25 var(--font-sans-1);}
+.kpi .delta{font:500 .75rem/1 var(--font-sans-2);color:#14772B;}
+@media (max-width:1024px){.kpi{grid-column:span 6;}}
+@media (max-width:640px){.kpi{grid-column:span 12;}}