… .claude/settings.local.json
… .github/workflows/green-scoreboard.yml
… 3_DAY_SURGICAL_PLAN.md
… ADR/ADR-001-Selector-Contract.md
… ADR/ADR-002-Feature-Flags-and-IA.md
… AI_AGENT_BACKTEST_FRAMEWORK.md
… BACKTEST_COMPLETE_SUMMARY.md
… BACKTEST_QUICKSTART.md
… BUILD_STRATEGY_EXECUTIVE_SUMMARY.md
… CAPITAL_ALLOCATION_FOLLOW_ON_TABLE.md
… CLAUDE_COOKBOOK_INTEGRATION.md
… COMPASS_DEMO_SLIDES.md
… COMPREHENSIVE_RESTRUCTURE_STRATEGY.md
… COOKBOOK_IMPLEMENTATION_SUMMARY.md
… DEMO_CHEAT_SHEET.md
… DEMO_ENHANCEMENTS.md
… DEMO_IMPLEMENTATION_COMPLETE.md
… DEMO_PREP_TOMORROW.md
… DEMO_QUICK_REFERENCE.md
… DEMO_QUICK_START.md
… DEMO_READY_SUMMARY.md
… DEMO_WALKTHROUGH.md
… Default Parameters/.env.example
… Default Parameters/ADR/ADR-001-Selector-Contract.md
… Default Parameters/ADR/ADR-002-Feature-Flags-and-IA.md
… Default Parameters/Capital Allocation Example.pdf
… Default Parameters/README.md
… Default Parameters/openapi/reserve-engine.yaml
… Default Parameters/src/core/selectors/__tests__/fundKpis.test.ts
… ENHANCED_DEMO_PLAN.md